OpenCV/requirements (2).txt
ultralytics
mediapipe
opencv-python
pyttsx3
Pillow
numpy